[UPDATE]Modification du type pour la 'note' donnée du commentaire

+++ src/main/java/wf3/project/alpha_betise/entities/CommentaireUtilisateur.java
@@ -27,7 +27,7 @@ public class CommentaireUtilisateur {
 	@NonNull
 	private Integer id;
 
-	private Integer note;
+	private Byte note;
 
 	private String commentaire;
